I wrote [last week][1] that "I don’t think there’s any tweaks to my current approach that will yield more than a 10x performance improvement." Then I had an insight! Inserting words into the EvalTree with the choices in board traversal order is a natural way to do it, but it is a choice. Rather than doing that, why not sort the choices to increase coherence?

This is a dramatic win. The initial bounds after building the tree drop by something like 5-20x. Lifting is less effective with these trees—the gap in bounds shrinks with every lift—but it never fully closes until the board is fully-lifted. This means that you can usually build the tree and go straight to `bound_remaining_boards` with no lifting, which saves a ton of memory. It's only for the very hardest boards where lifting is still helpful (and it is still extremely helpful).

- 50 board test set (3x4, 3 buckets): 333.2s → 40.1s, and only 700MB RAM peak.
- Board 234556: 275s → 319s (switchover 0), a loss! With switchover 4 it's only 24.4s, and switchover 5 is 17.5s.
- 20 board test set (4x4, 4 buckets): 1,014s → 236s (switchover=2) or 33s (mix).

Regardless, this is a huge win! And it should compound with letter grouping once #16 is fixed.

[1]: https://www.danvk.org/2025/02/13/boggle2025.html#:~:text=I%20don%E2%80%99t%20think%20there%E2%80%99s%20any%20tweaks%20to%20my%20current%20approach%20that%20will%20yield%20more%20than%20a%2010x%20performance%20improvement.